fix de codigo de service que estaba en el repo generico
